+++ apps/rails_app/app/javascript/packs/components.js
@@ -1,4 +1,0 @@
-import bootstrap from '../bootstrap'
-import App from '../components'
-
-bootstrap(App)