+++ apps/rails_app/app/javascript/packs/main.js
@@ -1,4 +1,0 @@
-import bootstrap from '../bootstrap'
-import App from '../main'
-
-bootstrap(App)