+++ libs/components/src/index.js
@@ -1,6 +1,3 @@
-import * as Utils from './utils'
-// export * from '@cwds/reactstrap'
-// from @cwds/reactstrap
 export {
   Badge,
   Breadcrumb,
@@ -56,35 +53,41 @@ export {
   UncontrolledDropdown,
   UncontrolledTooltip,
   Util,
-} from 'reactstrap'
-
-export { default as DataGrid } from '@cwds/datagrid'
-export { default as Rolodex } from './Rolodex'
-export { Alert, AlertUncontrolled } from './Alert'
-// export { default as AppBar } from './AppBar'
-export { default as Avatar } from './Avatar'
-export { default as Button } from './Button'
-export { default as Field } from './Field'
-export { Icon, ICON_NAMES } from '@cwds/icons'
-export { default as IconButton } from './IconButton'
-export { default as Infotip, UncontrolledInfotip } from './Infotip'
-export { default as InputMask } from './InputMask'
-export { default as JumpNav } from './JumpNav'
-export { default as JumpToTop } from './JumpToTop'
-export { Banner, Page, Footer } from './Layouts'
-export { default as PageTitle } from './PageTitle'
-export { default as PageActions } from './PageActions'
-export { default as Logo } from './Logo'
-// export { default as PageHeader } from './PageHeader'
-export { default as Placeholder } from './Placeholder'
-export { CollapsibleCard, UncontrolledCollapsibleCard } from './CollapsibleCard'
+} from '@cwds/reactstrap'
 export {
+  Rolodex,
+  Alert,
+  AlertUncontrolled,
+  Button,
+  Avatar,
+  Field,
+  IconButton,
+  Infotip,
+  UncontrolledInfotip,
+  InputMask,
+  JumpNav,
+  JumpToTop,
+  Banner,
+  Page,
+  Footer,
+  PageTitle,
+  PageActions,
+  Logo,
+  Select,
+  Placeholder,
+  CollapsibleCard,
+  UncontrolledCollapsibleCard,
+  Tile,
+  TileGroup,
+  TilePlaceholder,
+  defaultContext,
+  CaresProvider,
   Menu,
   UncontrolledMenu,
   MenuItem,
   UserMenu,
   UncontrolledUserMenu,
-} from './Menu'
-export { Utils }
-export { Tile, TileGroup, TilePlaceholder } from './Tile'
-export { defaultContext, CaresProvider } from './CaresContext'
+  Utils,
+} from '@cwds/cares'
+export { default as DataGrid } from '@cwds/datagrid'
+export { Icon, ICON_NAMES } from '@cwds/icons'